Create seperate components for all sections
