+++ apps/back-office/src/app/(dashboard)/settings/team-members/components/EditTeamMemberSidebar.tsx
@@ -21,7 +21,7 @@ export default function EditTeamMemberSidebar({
 }: EditTeamMemberSidebarProps) {
   return (
     <div
-      className={`fixed inset-y-0 right-0 w-96 bg-white shadow-lg transform transition-transform duration-300 ease-in-out z-50 ${
+      className={`fixed inset-y-0 right-0 w-[500px] bg-white shadow-xl transform transition-transform duration-300 ease-in-out z-50 ${
         isOpen ? "translate-x-0" : "translate-x-full"
       }`}
     >
